+++ CHANGELOG.md
@@ -19,8 +19,9 @@
 -   Template linting to CI using `djlint`
 -   Snapshot formatting check to CI using `djlint`
 -   Autoformatting of snapshots using `djlint`
--   Testing accross Django versions 2.2 - 4.0 and Python versions 3.8 - 3.11 using `tox`
+-   Testing across Django versions 2.2 - 4.0 and Python versions 3.8 - 3.11 using `tox`
 
 ### Changed
 
 -   Use snapshot testing plugin (syrupy) for component rendering tests instead of HTML fixtures
+-   form.helper (`FormHelper`) changed from a static `@property` to the form's `__init__` method to allow changes at runtime.